requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
streamlit==1.50.0
pandas==2.3.3
requests==2.32.5
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20
playwright==1.55.0
numpy==2.3.3

